refactor: introduce bite api package
