+++ libs/bite-tribe/api/src/lib/like-api.service.ts
@@ -2,7 +2,7 @@ import { ErrorHandler, inject, Injectable } from '@angular/core';
 import { AuthService } from 'ta-firestore';
 import { BehaviorSubject, skip, Subject } from 'rxjs';
 import { FirebaseFirestore } from '@capacitor-firebase/firestore';
-import { BITE_COLLECTION } from './bite-api.service';
+import { BITE_COLLECTION } from './bite-api/bite-api.service';
 import { User } from '@capacitor-firebase/authentication/dist/esm/definitions';
 
 const LIKES_COLLECTION_GROUP = 'likes';
