+++ libs/bite-tribe/api/src/lib/restaurant-api.service.ts
@@ -13,7 +13,7 @@ import {
 import { FirebaseFirestore } from '@capacitor-firebase/firestore';
 import { Link, Restaurant } from 'model';
 import { MENU_COLLECTION } from './menu-api.service';
-import { BITE_COLLECTION } from './bite-api.service';
+import { BITE_COLLECTION } from './bite-api/bite-api.service';
 
 const RESTAURANT_COLLECTION = 'restaurants';
